requirements.txt
azure-core==1.33.0
azure-cosmos==4.6.0
azure-storage-blob==12.25.1
blinker==1.8.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.2
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==2.1.5
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
scikit-learn==1.3.2
scipy==1.11.2
six==1.17.0
threadpoolctl==3.5.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
